// src/lib/types/steel/BeamCheckResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CheckStatus } from "./CheckStatus";
import type { CrossSectionClass } from "./CrossSectionClass";
import type { NamedCheck } from "./NamedCheck";

export type BeamCheckResult = { beam_id: number, profile_name: string, steel_grade: string, classification: CrossSectionClass, checks: Array<NamedCheck>, uc_max: number, status: CheckStatus, governing_check_id: string, };